Added folder/file: queries/GitHubQueries
